// src/interfaces/planta.interface.ts
// Generated by https://quicktype.io

export interface IPlant {
    id:               string;
    name:             string;
    binomialName:     string;
    price:            number;
    imgUrl:           string;
    wateringsPerWeek: number;
    fertilizerType:   string;
    heightInCm:       number;
}